Buggy state for radial enemies detection

+++ src/Grid.hpp
@@ -27,8 +27,8 @@ private:
 	unsigned int getXFromPos(float posX) { return (posX+m_gameAreaHWidth)/GRID_WIDTH; }
 	unsigned int getYFromPos(float posY) { return (posY+m_gameAreaHHeight)/GRID_HEIGHT; }
 	
-	float getPosFromX(unsigned int x) { return x*GRID_WIDTH-m_gameAreaHWidth; }
-	float getPosFromY(unsigned int y) { return y*GRID_HEIGHT-m_gameAreaHHeight; }
+	float getPosFromX(unsigned int x) { return static_cast<float>(x*GRID_WIDTH)-m_gameAreaHWidth; }
+	float getPosFromY(unsigned int y) { return static_cast<float>(y*GRID_HEIGHT)-m_gameAreaHHeight; }
 	
 	void addInZone(const Sphere& sphere, unsigned int x, unsigned int y, std::vector<Enemy*>& enemies);
